make a mlr model + model metadata for a specific task using a specific learner

makeModel(task, learner)

Arguments

task
an object of class mlr::makeRegrTask()

learner
an object of class mlr::makeLearner()

Value

a 2 elements named list : (1) snitch and (2) output. snitch is TRUE if function has provided the expected result. output is a named list which contains : (1) value : an object of class list made of 3 elements which are of classes "Model", "dataframe", and "list" (2) condition : a character specifying if the functions has encountered success, warning, error (3) message : the message relative to the condition

Examples

# NOT RUN {
# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T18:00:00Z",
  sensor = "tsa")

# extract the list of hourly sets of records
myDataset = myDataset$output$value

# create the tasks
myTasks = purrr::map(myDataset, makeTask, target = "tsa")

# extract the tasks from the outputs
myTasks = myTasks %>% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# keep the first task
myTask = myTasks[[1]]

# create the model
myModel = makeModel(
  task = myTask,
  learner = agrometeorLearners$mulLR_lonLatAlt_NA)

# extract the relevant information
myModel = myModel$output$value

# show the results of the LOOCV of this hour forthe choosen learner
head(myModel$predictions)
head(myModel$perfs$iters)

# show the model information
myModel$trained
# }