make a spatialization of a gridded dataset using a mlr model

make a spatialization of a gridded dataset using a mlr model

makeSpatialization(model, pred.grid = grid.df)

Arguments

model
an object of class mlr::train() that contains the prediction model

pred.grid
an object of class sf::st_makegrid(). This object must contains the same column names as the task on which the model has been trained

Value

A 2 elements named list

snitch : a boolean. Is TRUE if function has provided the expected result. Is FALSE is function throws an error

output : a named list which elements are :

value : a named list which elements are :

spatialized : an element of class data.frame. colnames are px (= reference of the pixel), response (= prediction value), se (= prediction standard error)

summary an element of class data.frame containing summary information about grid prediction. colnames are min.response, max.response, mean.response, min.se, max.se, mean.se

condition : a character specifying the condition encountered by the function: success, warning, or error.

message : a character specifying the message relative to the condition.

Examples

# NOT RUN {
# load magrittr for pipe use : %>%
library(magrittr)
# create the dataset
myDataset = makeDataset(
  dfrom = "2017-03-04T15:00:00Z",
  dto = "2017-03-04T18:00:00Z",
  sensor = "tsa")

# extract the list of hourly sets of records
myDataset = myDataset$output$value

# create the tasks
myTasks = purrr::map(myDataset, makeTask, target = "tsa")

# extract the tasks from the outputs
myTasks = myTasks %>% purrr::modify_depth(1, ~.$"output"$"value"$"task")

# keep the first task
myTask = myTasks[[1]]

# create the model
myModel = makeModel(
  task = myTask,
  learner = agrometeorLearners$mulLR_lonLatAlt_NA)

# extract the relevant information
myModel = myModel$output$value

# spatialize using the trained model
mySpatialization = makeSpatialization(
model = myModel$trained,
pred.grid = grid.df) # grid.df comes precompiled with the package

# get the relevant information
mySpatialization = mySpatialization$output$value

# show an excerpt of the spatialized data
head(mySpatialization$spatialized)

# show the summary stats of spatialized data
head(mySpatialization$summary)
# }