make a dataframe of stations records

makeDataset(user_token = Sys.getenv("AGROMET_API_V1_KEY"),
  stations = paste0(as.character(stations.df$sid), collapse = ","),
  json = NULL, dfrom = NULL, dto = NULL, sensor = "tsa",
  staticExpl = "elevation", dynExpl = NULL)

Arguments

stations
a character specifying the sid's of the stations to use separated by commas

json
a character specifying the path of a json file constructed from a batch db export. If NULL the API will be called.

dfrom
a datetime string specifying the dateTime Must have the form "YYYY-MM-DDTHH:MM:SSZ"

dto
a datetime string specifying the dateTime Must have the form "YYYY-MM-DDTHH:MM:SSZ"

sensor
a character specifying the sensor data you want to spatialize. One of tsa, hct, hra

staticExpl
a character vector specifying the static explanatory variables you want to add to the tasks. Any combinations of "altitude", "elevation", "slope", "aspect", "Agricultural_areas", "Artificials_surfaces", "Forest", "Herbaceous_vegetation". Latitude and longitude are always provided. Default = "Elevation"

dynExpl
a character vector specifying the dynamic explanatory variables you want to add to the task. Any combinations of inca, ens

token
a character specifying your agrometAPI token

Value

a list containing a boolean and a list of dataframes containing the desired records